Refactor of XMLFileMetaDataLoader in progress
